backend/api/requirements.txt
Flask~=3.1.0
Flask-Cors~=5.0.0
pillow~=11.2.1